Commit 5

+++ Ejercicio 3/funciones.h
@@ -5,8 +5,8 @@
 #ifndef FUNCIONES_H
 #define FUNCIONES_H
 
-float suma(float nota_primera, float nota_segunda, float nota_tercera, float nota_cuarta, float nota_quinta, float nota_sexta, float nota_septima, float nota_octava);
-float promedio(float nota_primera, float nota_segunda, float nota_tercera, float nota_cuarta, float nota_quinta, float nota_sexta, float nota_septima, float nota_octava);
-
+double suma(double nota_primera, double nota_segunda, double nota_tercera, double nota_cuarta, double nota_quinta, double nota_sexta, double nota_septima, double nota_octava);
+double promedio(double nota_primera, double nota_segunda, double nota_tercera, double nota_cuarta, double nota_quinta, double nota_sexta, double nota_septima, double nota_octava);
+//double promedio2(double suma);
 
 #endif //FUNCIONES_H
